requirements.txt
selenium~=3.141.0
python-dotenv~=0.15.0
urllib3==1.26.2
mysql-connector-python~=8.0.23
chromedriver-autoinstaller
docker~=4.4.4
python-dateutil~=2.8.1
setuptools~=53.0.0